+++ Makefile
@@ -12,7 +12,8 @@ SRCS = ${SRCDIR}/json.hpp \
 			 ${SRCDIR}/detail/conversions/to_json.hpp \
 			 ${SRCDIR}/detail/parsing/input_adapters.hpp \
 			 ${SRCDIR}/detail/parsing/lexer.hpp \
-			 ${SRCDIR}/detail/parsing/parser.hpp
+			 ${SRCDIR}/detail/parsing/parser.hpp \
+			 ${SRCDIR}/detail/iterators/primitive_iterator.hpp
